Update README.md
My intent was to convert the warning and info blocks to the native markdown options which is (among other things) mobile friendly (https://github.com/orgs/community/discussions/16925). Unfortunately, this did not seem to work as expected, so I used the next best thing. 

I also made some other improvements. The changes to the chapter layout are with the intention to more easily add additional content.