Check if luxtronik_key exists in luxtronik coordinator before adding entity
1) Check if luxtronik_key exists in luxtronik coordinator before adding entity
Previously, we used the luxtronik firmware version to determine if an entity should be added or not. This was complex, because we did not always know exactly at which FW a parameter was added. Recently, #461 and #439 are contradictory: a parameter exists in V2.88.3 but not in V1.88.3  ?!
To prevent this mess, the integration now checks if the entity key exists in the Luxtronik data. If not, the integration will automatically skip adding it.
So no more need for firmware versions checks!

2) Some common code parts of multiple platforms moved to base.py

3) Added Debouncer to numbers, water heater and climte, to prevent high-frequent writing changed values to heatpunp.
Removed debouncer.py (was not used)

4) Improved startup-behavior, so entities should always show data immedately